Added My Homework 2 Via Upload
In this project I have coded a webpage that takes in user input for either Celsius or Fahrenheit and converts it into the opposite measurement (Example: Fahrenheit --> Celsius). The webpage only allows for one input to be taken and then displays a visual image of the weather that depends on the temperature. (Example: 70 degrees Fahrenheit displays a picture of the sun)